Fix problem on systems where size_t is only 32-bits

+++ ADApp/pluginSrc/NDFileHDF5Dataset.cpp
@@ -6,7 +6,7 @@
 #include <hdf5_hl.h>
 
 #ifndef htonll
-#define htonll(x) ( ( (uint64_t)(htonl( (uint32_t)((x << 32) >> 32)))<< 32) | htonl( ((uint32_t)(x >> 32)) ))
+#define htonll(x) ( ( (uint64_t)(htonl( (uint32_t)(((uint64_t)x << 32) >> 32)))<< 32) | htonl( ((uint32_t)((uint64_t)x >> 32)) ))
 #endif
 
 static const char *fileName = "NDFileHDF5Dataset";
